Commented out code that needs and some to be fixed. Plus some revision daily backend
I've removed the packages that were install causing the to compile error from package.json (to reinstall need to do new 'npm i <packageName>'), so need to check the packages that were installed to make see what is going wrong. Also commented out the sleep page that made use of the libraries.

Changes to revision daily page:
 - Able to select individual pomodoro items
 - Able to delete pomodoro items
 - Somewhat able to complete one pomodoro cycle.

To fix in revision daily page:
 - unable to do more than 1 pomodoro  (25 minutes) of each item, unless refresh page.

To add:
 - long break after goal is reached,
 - delete once all progress finished? and add to point system?