Updates to overview comp. UI +  changes to sleep page
Made the overview display a graph.

The sleep page now gets data and from every day of the week, to display on the graph
Also, added a goal to the sleep supabase table, the point of this is to show what the current goal at the time was when the sleep data was added, to be displayed on the graph.

TODO:
 - Get data to display on overview graph
 - Get other data to relate to overview, i.e avg. sleep time
 - Add the same thing done above to water.
 - And more.....