Graphs now work in sleep and water pages.
I've made a new file to hold the graph components in the website.

Within the new component file, 'Graphs.js', I've made a new AreaGraph component, this will handle to area graphs that display data for a week.

I've also added a amount_goal to  the water table in the DB, so it can be displayed on the graph. Like I did with water in my last commit.