Fix crashes and changes to dashboard UI
Just removed the buttons and replaced them with overview component.

FIX:
Fixed the crash that would occur when adding new data to the db.
When the AreaGraph component received a data value of null, it would crash, since you can't get the length of null or pass it into the Area Chart component from recharts.

TODO:
 - Finish overview component
 - Decide what data we are going to compare and how to present it on the UI.